refact: split apart login schema and page
