+++ if-food-front/src/pages/login/schema.ts
@@ -1,0 +1,14 @@
+import * as yup from "yup";
+
+export interface LoginFormData {
+    email: string;
+    password: string;
+}
+
+export const schema = yup.object().shape({
+    email: yup.string().email("Email inválido").required("Email é obrigatório"),
+    password: yup
+        .string()
+        .min(6, "A senha deve ter pelo menos 6 caracteres")
+        .required("Senha é obrigatória"),
+});
