+++ if-food-front/src/routes.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter, Outlet } from "react-router-dom";
-import { LoginPage } from "./pages/login-page";
+import { LoginPage } from "./pages/login";
 import { RegisterPage } from "./pages/register-page";
 import { HomePage } from "./pages/home-page";
 import { Guest } from "./components/guest";
